Move pipelines to radiology app
