+++ sample-apps/radiology/lib/infers/vertebra_pipeline.py
